using ErrorBoundary to catch all errors
